% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_args.R
\name{process_args}
\alias{process_args}
\title{Convert command arguments into a named list.}
\usage{
process_args(
  args,
  required_args,
  optional_value_args,
  optional_flag_args,
  ignored_args = c("no-echo", "no-restore", "file", "args"),
  quietly = FALSE
)
}
\arguments{
\item{args}{Character vector in form of `--<parameter>`  or
`--<parameter>=<value>`, usually derived by calling commandArgs()}

\item{optional_value_args}{A character vector of parameter names that are not
required.  When present, these parameters may have, but do not require,
argument values.}

\item{ignored_args}{A character vector of parameter names that will be
removed from the return value.}

\item{quietly}{Set to TRUE to suppress printing of arguments to console even
if `args` contains '--verbose'}

\item{required_value_args}{A character vector of parameter names that are
required on the command line and which must be assigned arguments values.}

\item{required_flag_args}{A character vector of parameter names that are
required on the command line; these parameters may have, but do not require
argument values.}
}
\value{
A named character vector where the names are parameters names.  The
  values are "TRUE" if the parameter names was present, "FALSE" if the
  parameter name was absent, or the value after the `=` if assigned a value
  on in `args`.

  If `args` contains a value `--verbose` the return vector is printed
  to the console unless `quietly` is set to TRUE.
}
\description{
Converts results from commandArgs() into a named character vector where the
names are parameter names and the values are the argument values from the
command line.
}
\details{
`process_args` needs to know all allowable parameter names from the command
line.  Given the names of allowable parameters, the function:

\itemize{
  \item Checks that all parameter=arguments pairs are formatted correctly
  (`--<parameter>=<argument>` or `--<parameter>`);
  \item Checks to be sure that all required parameter names are present;
  \item Checks to all parameters that need argument values have them;
  \item Checks to be sure that no unrecognized parameters exist;
  \item Inserts any missing parameters, assigning an argument value of "FALSE"
  \item Strips out the ignored parameters;
  \item Returns the complete parameter-argument pair values as a character
  vector of argument values named by parameter.
}
}
